+++ neo4j/_async/work/transaction.py
@@ -16,6 +16,8 @@
 # limitations under the License.
 
 
+from functools import wraps
+
 from ..._async_compat.util import AsyncUtil
 from ...data import DataHydrator
 from ...exceptions import TransactionError
@@ -24,17 +26,10 @@ from ..io import ConnectionErrorHandler
 from .result import AsyncResult
 
 
-class AsyncTransaction:
-    """ Container for multiple Cypher queries to be executed within a single
-    context. asynctransactions can be used within a :py:const:`async with`
-    block where the transaction is committed or rolled back on based on
-    whether an exception is raised::
+__all__ = ("AsyncTransaction", "AsyncManagedTransaction")
 
-        async with session.begin_transaction() as tx:
-            ...
 
-    """
-
+class _AsyncTransactionBase:
     def __init__(self, connection, fetch_size, on_closed, on_error):
         self._connection = connection
         self._error_handling_connection = ConnectionErrorHandler(
@@ -42,22 +37,22 @@ class AsyncTransaction:
         )
         self._bookmark = None
         self._results = []
-        self._closed = False
+        self._closed_flag = False
         self._last_error = None
         self._fetch_size = fetch_size
         self._on_closed = on_closed
         self._on_error = on_error
 
-    async def __aenter__(self):
+    async def _enter(self):
         return self
 
-    async def __aexit__(self, exception_type, exception_value, traceback):
-        if self._closed:
+    async def _exit(self, exception_type, exception_value, traceback):
+        if self._closed_flag:
             return
         success = not bool(exception_type)
         if success:
-            await self.commit()
-        await self.close()
+            await self._commit()
+        await self._close()
 
     async def _begin(
         self, database, imp_user, bookmarks, access_mode, metadata, timeout
@@ -105,14 +100,16 @@ class AsyncTransaction:
         :param parameters: dictionary of parameters
         :type parameters: dict
         :param kwparameters: additional keyword parameters
-        :returns: a new :class:`neo4j.Result` object
-        :rtype: :class:`neo4j.Result`
+
+        :returns: a new :class:`neo4j.AsyncResult` object
+        :rtype: :class:`neo4j.AsyncResult`
+
         :raise TransactionError: if the transaction is already closed
         """
         if isinstance(query, Query):
             raise ValueError("Query object is only supported for session.run")
 
-        if self._closed:
+        if self._closed_flag:
             raise TransactionError(self, "Transaction closed")
         if self._last_error:
             raise TransactionError(self,
@@ -136,12 +133,12 @@ class AsyncTransaction:
 
         return result
 
-    async def commit(self):
+    async def _commit(self):
         """Mark this transaction as successful and close in order to trigger a COMMIT.
 
         :raise TransactionError: if the transaction is already closed
         """
-        if self._closed:
+        if self._closed_flag:
             raise TransactionError(self, "Transaction closed")
         if self._last_error:
             raise TransactionError(self,
@@ -156,17 +153,17 @@ class AsyncTransaction:
             await self._connection.fetch_all()
             self._bookmark = metadata.get("bookmark")
         finally:
-            self._closed = True
+            self._closed_flag = True
             await AsyncUtil.callback(self._on_closed)
 
         return self._bookmark
 
-    async def rollback(self):
+    async def _rollback(self):
         """Mark this transaction as unsuccessful and close in order to trigger a ROLLBACK.
 
         :raise TransactionError: if the transaction is already closed
         """
-        if self._closed:
+        if self._closed_flag:
             raise TransactionError(self, "Transaction closed")
 
         metadata = {}
@@ -180,20 +177,82 @@ class AsyncTransaction:
                 await self._connection.send_all()
                 await self._connection.fetch_all()
         finally:
-            self._closed = True
+            self._closed_flag = True
             await AsyncUtil.callback(self._on_closed)
 
-    async def close(self):
+    async def _close(self):
         """Close this transaction, triggering a ROLLBACK if not closed.
         """
-        if self._closed:
+        if self._closed_flag:
             return
-        await self.rollback()
+        await self._rollback()
 
-    def closed(self):
+    def _closed(self):
         """Indicator to show whether the transaction has been closed.
 
         :return: :const:`True` if closed, :const:`False` otherwise.
         :rtype: bool
         """
-        return self._closed
+        return self._closed_flag
+
+
+class AsyncTransaction(_AsyncTransactionBase):
+    """ Container for multiple Cypher queries to be executed within a single
+    context. asynctransactions can be used within a :py:const:`async with`
+    block where the transaction is committed or rolled back on based on
+    whether an exception is raised::
+
+        async with session.begin_transaction() as tx:
+            ...
+
+    """
+
+    @wraps(_AsyncTransactionBase._enter)
+    async def __aenter__(self):
+        return await self._enter()
+
+    @wraps(_AsyncTransactionBase._exit)
+    async def __aexit__(self, exception_type, exception_value, traceback):
+        await self._exit(exception_type, exception_value, traceback)
+
+    @wraps(_AsyncTransactionBase._commit)
+    async def commit(self):
+        return await self._commit()
+
+    @wraps(_AsyncTransactionBase._rollback)
+    async def rollback(self):
+        return await self._rollback()
+
+    @wraps(_AsyncTransactionBase._close)
+    async def close(self):
+        return await self._close()
+
+    @wraps(_AsyncTransactionBase._closed)
+    def closed(self):
+        return self._closed()
+
+
+class AsyncManagedTransaction(_AsyncTransactionBase):
+    """Transaction object provided to transaction functions.
+
+    Inside a transaction function, the driver is responsible for managing
+    (committing / rolling back) the transaction. Therefore,
+    AsyncManagedTransactions don't offer such methods.
+    Otherwise, they behave like :class:`.AsyncTransaction`.
+
+    * To commit the transaction,
+      return anything from the transaction function.
+    * To rollback the transaction, raise any exception.
+
+    Note that transaction functions have to be idempontent (i.e., the result
+    of running the function once has to be the same as running it any number
+    of times). This is, because the driver will retry the transaction function
+    if the error is classified as retriable.
+
+    .. versionadded:: 5.0
+
+        Prior, transaction functions used :class:`AsyncTransaction` objects,
+        but would cause hard to interpret errors when managed explicitly
+        (committed or rolled back by user code).
+    """
+    pass