+++ tests/unit/async_/io/test_class_bolt3.py
@@ -22,10 +22,7 @@ from neo4j._async.io._bolt3 import AsyncBolt3
 from neo4j.conf import PoolConfig
 from neo4j.exceptions import ConfigurationError
 
-from ...._async_compat import (
-    AsyncMagicMock,
-    mark_async_test,
-)
+from ...._async_compat import mark_async_test
 
 
 @pytest.mark.parametrize("set_stale", (True, False))
@@ -99,11 +96,11 @@ async def test_simple_pull(fake_socket):
 @pytest.mark.parametrize("recv_timeout", (1, -1))
 @mark_async_test
 async def test_hint_recv_timeout_seconds_gets_ignored(
-    fake_socket_pair, recv_timeout
+    fake_socket_pair, recv_timeout, mocker
 ):
     address = ("127.0.0.1", 7687)
     sockets = fake_socket_pair(address)
-    sockets.client.settimeout = AsyncMagicMock()
+    sockets.client.settimeout = mocker.AsyncMock()
     await sockets.server.send_message(0x70, {
         "server": "Neo4j/3.5.0",
         "hints": {"connection.recv_timeout_seconds": recv_timeout},