+++ tests/unit/async_/work/conftest.py
@@ -1,0 +1,4 @@
+from ._fake_connection import (
+    async_fake_connection,
+    async_fake_connection_generator,
+)