+++ tests/unit/sync/io/test_neo4j_pool.py
@@ -16,8 +16,6 @@
 # limitations under the License.
 
 
-from unittest.mock import Mock
-
 import pytest
 
 from neo4j import (
@@ -36,11 +34,8 @@ from neo4j.exceptions import (
     SessionExpired,
 )
 
-from ...._async_compat import (
-    mark_sync_test,
-    Mock,
-)
-from ..work import FakeConnection
+from ...._async_compat import mark_sync_test
+from ..work import fake_connection_generator
 
 
 ROUTER_ADDRESS = ResolvedAddress(("1.2.3.1", 9001), host_name="host")
@@ -49,12 +44,12 @@ WRITER_ADDRESS = ResolvedAddress(("1.2.3.1", 9003), host_name="host")
 
 
 @pytest.fixture()
-def opener():
+def opener(fake_connection_generator, mocker):
     def open_(addr, timeout):
-        connection = FakeConnection()
+        connection = fake_connection_generator()
         connection.addr = addr
         connection.timeout = timeout
-        route_mock = Mock()
+        route_mock = mocker.Mock()
         route_mock.return_value = [{
             "ttl": 1000,
             "servers": [
@@ -67,7 +62,7 @@ def opener():
         opener_.connections.append(connection)
         return connection
 
-    opener_ = Mock()
+    opener_ = mocker.Mock()
     opener_.connections = []
     opener_.side_effect = open_
     return opener_