+++ tests/unit/sync/work/test_transaction.py
@@ -26,7 +26,7 @@ from neo4j import (
     Transaction,
 )
 
-from ._fake_connection import fake_connection
+from ...._async_compat import mark_sync_test
 
 
 @pytest.mark.parametrize(("explicit_commit", "close"), (
@@ -34,13 +34,15 @@ from ._fake_connection import fake_connection
     (True, False),
     (True, True),
 ))
-def test_transaction_context_when_committing(mocker, fake_connection,
-                                             explicit_commit, close):
-    on_closed = MagicMock()
-    on_error = MagicMock()
+@mark_sync_test
+def test_transaction_context_when_committing(
+    mocker, fake_connection, explicit_commit, close
+):
+    on_closed = mocker.Mock()
+    on_error = mocker.Mock()
     tx = Transaction(fake_connection, 2, on_closed, on_error)
-    mock_commit = mocker.patch.object(tx, "commit", wraps=tx.commit)
-    mock_rollback = mocker.patch.object(tx, "rollback", wraps=tx.rollback)
+    mock_commit = mocker.patch.object(tx, "_commit", wraps=tx._commit)
+    mock_rollback = mocker.patch.object(tx, "_rollback", wraps=tx._rollback)
     with tx as tx_:
         assert mock_commit.call_count == 0
         assert mock_rollback.call_count == 0
@@ -48,7 +50,7 @@ def test_transaction_context_when_committing(mocker, fake_connection,
         if explicit_commit:
             tx_.commit()
             mock_commit.assert_called_once_with()
-            assert tx.closed()
+            assert tx_.closed()
         if close:
             tx_.close()
             assert tx_.closed()
@@ -62,13 +64,15 @@ def test_transaction_context_when_committing(mocker, fake_connection,
     (False, True),
     (True, True),
 ))
-def test_transaction_context_with_explicit_rollback(mocker, fake_connection,
-                                                    rollback, close):
-    on_closed = MagicMock()
-    on_error = MagicMock()
+@mark_sync_test
+def test_transaction_context_with_explicit_rollback(
+    mocker, fake_connection, rollback, close
+):
+    on_closed = mocker.Mock()
+    on_error = mocker.Mock()
     tx = Transaction(fake_connection, 2, on_closed, on_error)
-    mock_commit = mocker.patch.object(tx, "commit", wraps=tx.commit)
-    mock_rollback = mocker.patch.object(tx, "rollback", wraps=tx.rollback)
+    mock_commit = mocker.patch.object(tx, "_commit", wraps=tx._commit)
+    mock_rollback = mocker.patch.object(tx, "_rollback", wraps=tx._rollback)
     with tx as tx_:
         assert mock_commit.call_count == 0
         assert mock_rollback.call_count == 0
@@ -86,15 +90,18 @@ def test_transaction_context_with_explicit_rollback(mocker, fake_connection,
     assert tx_.closed()
 
 
-def test_transaction_context_calls_rollback_on_error(mocker, fake_connection):
+@mark_sync_test
+def test_transaction_context_calls_rollback_on_error(
+    mocker, fake_connection
+):
     class OopsError(RuntimeError):
         pass
 
     on_closed = MagicMock()
     on_error = MagicMock()
     tx = Transaction(fake_connection, 2, on_closed, on_error)
-    mock_commit = mocker.patch.object(tx, "commit", wraps=tx.commit)
-    mock_rollback = mocker.patch.object(tx, "rollback", wraps=tx.rollback)
+    mock_commit = mocker.patch.object(tx, "_commit", wraps=tx._commit)
+    mock_rollback = mocker.patch.object(tx, "_rollback", wraps=tx._rollback)
     with pytest.raises(OopsError):
         with tx as tx_:
             assert mock_commit.call_count == 0
@@ -112,8 +119,10 @@ def test_transaction_context_calls_rollback_on_error(mocker, fake_connection):
     ({"x": {(1, 2): '1+2i', (2, 0): '2'}}, TypeError),
     ({"x": uuid4()}, TypeError),
 ))
-def test_transaction_run_with_invalid_parameters(fake_connection, parameters,
-                                                 error_type):
+@mark_sync_test
+def test_transaction_run_with_invalid_parameters(
+    fake_connection, parameters, error_type
+):
     on_closed = MagicMock()
     on_error = MagicMock()
     tx = Transaction(fake_connection, 2, on_closed, on_error)
@@ -121,6 +130,7 @@ def test_transaction_run_with_invalid_parameters(fake_connection, parameters,
         tx.run("RETURN $x", **parameters)
 
 
+@mark_sync_test
 def test_transaction_run_takes_no_query_object(fake_connection):
     on_closed = MagicMock()
     on_error = MagicMock()
@@ -129,10 +139,14 @@ def test_transaction_run_takes_no_query_object(fake_connection):
         tx.run(Query("RETURN 1"))
 
 
-def test_transaction_rollbacks_on_open_connections(fake_connection):
-    tx = Transaction(fake_connection, 2,
-                     lambda *args, **kwargs: None,
-                     lambda *args, **kwargs: None)
+@mark_sync_test
+def test_transaction_rollbacks_on_open_connections(
+    fake_connection
+):
+    tx = Transaction(
+        fake_connection, 2, lambda *args, **kwargs: None,
+        lambda *args, **kwargs: None
+    )
     with tx as tx_:
         fake_connection.is_reset_mock.return_value = False
         fake_connection.is_reset_mock.reset_mock()
@@ -142,44 +156,56 @@ def test_transaction_rollbacks_on_open_connections(fake_connection):
         fake_connection.rollback.assert_called_once()
 
 
-def test_transaction_no_rollback_on_reset_connections(fake_connection):
-    tx = Transaction(fake_connection, 2,
-                     lambda *args, **kwargs: None,
-                     lambda *args, **kwargs: None)
+@mark_sync_test
+def test_transaction_no_rollback_on_reset_connections(
+    fake_connection
+):
+    tx = Transaction(
+        fake_connection, 2, lambda *args, **kwargs: None,
+        lambda *args, **kwargs: None
+    )
     with tx as tx_:
         fake_connection.is_reset_mock.return_value = True
         fake_connection.is_reset_mock.reset_mock()
         tx_.rollback()
         fake_connection.is_reset_mock.assert_called_once()
-        fake_connection.reset.asset_not_called()
-        fake_connection.rollback.asset_not_called()
+        fake_connection.reset.assert_not_called()
+        fake_connection.rollback.assert_not_called()
 
 
-def test_transaction_no_rollback_on_closed_connections(fake_connection):
-    tx = Transaction(fake_connection, 2,
-                     lambda *args, **kwargs: None,
-                     lambda *args, **kwargs: None)
+@mark_sync_test
+def test_transaction_no_rollback_on_closed_connections(
+    fake_connection
+):
+    tx = Transaction(
+        fake_connection, 2, lambda *args, **kwargs: None,
+        lambda *args, **kwargs: None
+    )
     with tx as tx_:
         fake_connection.closed.return_value = True
         fake_connection.closed.reset_mock()
         fake_connection.is_reset_mock.reset_mock()
         tx_.rollback()
         fake_connection.closed.assert_called_once()
-        fake_connection.is_reset_mock.asset_not_called()
-        fake_connection.reset.asset_not_called()
-        fake_connection.rollback.asset_not_called()
+        fake_connection.is_reset_mock.assert_not_called()
+        fake_connection.reset.assert_not_called()
+        fake_connection.rollback.assert_not_called()
 
 
-def test_transaction_no_rollback_on_defunct_connections(fake_connection):
-    tx = Transaction(fake_connection, 2,
-                     lambda *args, **kwargs: None,
-                     lambda *args, **kwargs: None)
+@mark_sync_test
+def test_transaction_no_rollback_on_defunct_connections(
+    fake_connection
+):
+    tx = Transaction(
+        fake_connection, 2, lambda *args, **kwargs: None,
+        lambda *args, **kwargs: None
+    )
     with tx as tx_:
         fake_connection.defunct.return_value = True
         fake_connection.defunct.reset_mock()
         fake_connection.is_reset_mock.reset_mock()
         tx_.rollback()
         fake_connection.defunct.assert_called_once()
-        fake_connection.is_reset_mock.asset_not_called()
-        fake_connection.reset.asset_not_called()
-        fake_connection.rollback.asset_not_called()
+        fake_connection.is_reset_mock.assert_not_called()
+        fake_connection.reset.assert_not_called()
+        fake_connection.rollback.assert_not_called()